feat: modal button linking
